# Reconnect sql when current connection has error