--- a/src/main/java/com/yugabyte/sample/apps/SqlSecondaryIndex.java
+++ b/src/main/java/com/yugabyte/sample/apps/SqlSecondaryIndex.java
@@ -143,7 +143,8 @@ public class SqlSecondaryIndex extends AppBase {
         }
       }
     } catch (Exception e) {
-      LOG.fatal("Failed reading value: " + key.getValueStr(), e);
+      LOG.info("Failed reading value: " + key.getValueStr(), e);
+      preparedSelect = null;
       return 0;
     }
     return 1;
@@ -176,7 +177,8 @@ public class SqlSecondaryIndex extends AppBase {
       getSimpleLoadGenerator().recordWriteSuccess(key);
     } catch (Exception e) {
       getSimpleLoadGenerator().recordWriteFailure(key);
-      LOG.fatal("Failed writing key: " + key.asString(), e);
+      LOG.info("Failed writing key: " + key.asString(), e);
+      preparedInsert = null;
     }
     return result;
   }