Operations in Sql classes such as SqlInserts would keep retrying upon connection errors.

Start SqlInserts workload of the sample app against a running local cluster, then stop some TS process (which the sample app connects to) in the local cluster (the cluster as a whole is still available), and the sample app resumes normal operations by getting connection to running TS.