@nchandrappa 2020-12-09
@tedyu This is not a valid testcase. 

 ``` Started SqlInserts workload of the sample app against a stopped local cluster (leading to sample app waiting for the cluster), then started the local cluster, and the sample app resuming normal operations. ``` 

We need to test with a healthy cluster then we'll need to bring down a set of nodes in the cluster such that the cluster is still available and verify that workload throughput doesn't go to 0.