@tedyu 2020-12-09
Tested on a local cluster with 3 TS.
Once the previous TS with connection was killed, I saw the following in sample app output:
```
36471 [Thread-2] INFO com.yugabyte.sample.apps.SqlInserts  - Wrote key: bd75a8a1-0a3e-4d95-9508-a6c57859b5c0:14, val:14, return code: 1
36525 [Thread-3] INFO com.yugabyte.sample.apps.SqlInserts  - Wrote key: bd75a8a1-0a3e-4d95-9508-a6c57859b5c0:15, val:15, return code: 1
36997 [Thread-4] FATAL com.yugabyte.sample.apps.SqlInserts  - Failed reading value: val:11
org.postgresql.util.PSQLException: FATAL: terminating connection due to administrator command
    at org.postgresql.core.v3.QueryExecutorImpl.receiveErrorResponse(QueryExecutorImpl.java:2553)
    at org.postgresql.core.v3.QueryExecutorImpl.processResults(QueryExecutorImpl.java:2285)
    at org.postgresql.core.v3.QueryExecutorImpl.execute(QueryExecutorImpl.java:323)
    at org.postgresql.jdbc.PgStatement.executeInternal(PgStatement.java:473)
    at org.postgresql.jdbc.PgStatement.execute(PgStatement.java:393)
    at org.postgresql.jdbc.PgPreparedStatement.executeWithFlags(PgPreparedStatement.java:164)
    at org.postgresql.jdbc.PgPreparedStatement.executeQuery(PgPreparedStatement.java:114)
    at com.yugabyte.sample.apps.SqlInserts.doRead(SqlInserts.java:120)
    at com.yugabyte.sample.apps.AppBase.performRead(AppBase.java:767)
    at com.yugabyte.sample.common.IOPSThread.run(IOPSThread.java:88)
37141 [Thread-4] INFO com.yugabyte.sample.apps.AppBase  - Connected to 127.0.0.1:5433
40609 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 0.40 ops/sec (5002.00 ms/op), 14 total ops  |  Write: 0.40 ops/sec (5004.50 ms/op), 14 total ops  |  Uptime: 40010 ms |
```
There was only one exception. After which the connection was re-established and app proceeded.