@nchandrappa 2020-12-10
@tedyu Awesome. LGTM.

I did a test on Dev portal with 5-node YB cluster, brought down 2 nodes in an AZ to verify the behavior. Throughput didn't got to 0 and normalized.

Before:

```
331158 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 5262.71 ops/sec (1.52 ms/op), 1729334 total ops  |  Write: 720.51 ops/sec (2.77 ms/op), 238599 total ops  |  Uptime: 330053 ms |
336159 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 5301.16 ops/sec (1.51 ms/op), 1755842 total ops  |  Write: 733.73 ops/sec (2.73 ms/op), 242268 total ops  |  Uptime: 335054 ms |
341162 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 5307.58 ops/sec (1.51 ms/op), 1782383 total ops  |  Write: 734.92 ops/sec (2.72 ms/op), 245943 total ops  |  Uptime: 340057 ms |
346163 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 5327.25 ops/sec (1.50 ms/op), 1809039 total ops  |  Write: 731.65 ops/sec (2.73 ms/op), 249604 total ops  |  Uptime: 345058 ms |
```
after:
```
1411271 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 4340.61 ops/sec (1.84 ms/op), 5591825 total ops  |  Write: 478.35 ops/sec (4.18 ms/op), 698078 total ops  |  Uptime: 1410166 ms |
1416271 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 4352.14 ops/sec (1.84 ms/op), 5613588 total ops  |  Write: 478.95 ops/sec (4.18 ms/op), 700473 total ops  |  Uptime: 1415166 ms |
1421272 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 4358.53 ops/sec (1.83 ms/op), 5635382 total ops  |  Write: 479.57 ops/sec (4.17 ms/op), 702871 total ops  |  Uptime: 1420167 ms |
1426272 [Thread-1] INFO com.yugabyte.sample.common.metrics.MetricsTracker  - Read: 4342.24 ops/sec (1.84 ms/op), 5657095 total ops  |  Write: 481.96 ops/sec (4.15 ms/op), 705281 total ops  |  Uptime: 1425167 ms |
```